script further developed

+++ installDefiMasternode.sh
@@ -10,16 +10,18 @@ trap cleanup 0 SIGHUP SIGINT SIGQUIT SIGABRT SIGTERM EXIT
 
 #_variables____________________________________________________________________
 
-__dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
-__file="${__dir}/$(basename "${BASH_SOURCE[0]}")"
-__base="$(basename ${__file} .sh)"
-__root="$(cd "$(dirname "${__dir}")" && pwd)" 
+#__dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+#__file="${__dir}/$(basename "${BASH_SOURCE[0]}")"
+#__base="$(basename ${__file} .sh)"
+#__root="$(cd "$(dirname "${__dir}")" && pwd)"
+
+USERNAME="defichain"
 
 #_constants____________________________________________________________________
-SUCCESS=0
-FILE_NOT_FOUND=240
-INSTALLATION_ABORTED=300
-NO_ROOT=1
+
+EXIT_CODE_NO_ROOT=3
+EXIT_CODE_INSTALLATION_ABORTED=4
+
 
 RED='\033[0;31m'
 GREEN='\033[0;32m'
@@ -30,23 +32,36 @@ INFO=${INFO:-"1"}
 DEBUG=${DEBUG:-"0"}
 ERROR=${ERROR:-"1"}
 WARNING=${WARNING:-"1"}
+SUCCESS=${SUCCESS:-"1"}
 
+#_DEFI constants_______________________________________________________________
 MASTERNODE_ARCHIVE="defichain-1.8.2-x86_64-pc-linux-gnu.tar.gz"
 MASTERNODE_DOWNLOAD_URL="https://github.com/DeFiCh/ain/releases/download/v1.8.2/$MASTERNODE_ARCHIVE"
 MASTERNODE_RELEASE_FOLDER="defichain-1.8.2"
-MASTERNODE_TARGET_FOLDER="$HOME/.defi"
+
+
+SNAPSHOT_DOWNLOAD_URL="https://defi-snapshots-europe.s3.eu-central-1.amazonaws.com/snapshot-mainnet-1052243.zip"
 
 #_functions____________________________________________________________________
 
 function cleanup() {
-  # for eg. rm -f "/tmp/${lock_file}.lock"
-  rm -f /tmp/$MASTERNODE_ARCHIVE
-  rm -r -f /tmp/$MASTERNODE_RELEASE_FOLDER
+
+    if [ $? -ne $EXIT_CODE_NO_ROOT ]
+        then
+            __msg_info "Start cleanup"
+
+            # -f "/tmp/${lock_file}.lock"
+            rm -f /tmp/$MASTERNODE_ARCHIVE
+            rm -r -f /tmp/$MASTERNODE_RELEASE_FOLDER
+
+            __msg_success "Cleanup finished"
+    fi
+   
 }
 
-# Liest einen Wert von der Tastatur ein und gibt diesen zurück
-# Aufruf mit Defaultvalue: variable=$(promptValue "Text der Abfrage" $defaultwert)
-# Aufruf ohne Defaultvalue: variable=$(promptValue "Text der Abfrage")
+# Reads a value from the keyboard and returns it
+# Call with defaultvalue: variable=$(promptValue "text of the query" $defaultvalue)
+# Call without default value: variable=$(promptValue "Text of the query")
 function promptValue() {
  if [ -z ${2+x} ]; 
  	then 
@@ -55,64 +70,134 @@ function promptValue() {
 		read -r -e -p "$1 [$2]:" val
 		val=${val:-$2}
  	fi
- echo $val
+ echo "$val"
 }
 
 function __msg_error() {
-    [[ "${ERROR}" == "1" ]] && echo -e "${RED}[ERROR]: $*" && echo -e "${NC}"
+    [[ "${ERROR}" == "1" ]] && echo -e "${RED}[ERROR]:   $*" && echo -ne "${NC}"
 }
 
 function __msg_warning() {
-    [[ "${WARNING}" == "1" ]] && echo -e "${YELLOW}[WARNING]: $*" && echo -e "${NC}"
+    [[ "${WARNING}" == "1" ]] && echo -e "${YELLOW}[WARNING]: $*" && echo -ne "${NC}"
+}
+
+function __msg_success() {
+    [[ "${SUCCESS}" == "1" ]] && echo -e "${GREEN}[SUCCESS]: $*" && echo -ne "${NC}"
 }
 
 function __msg_info() {
-    [[ "${INFO}" == "1" ]] && echo -e "${GREEN}[INFO]: $*" && echo -e "${NC}"
+    [[ "${INFO}" == "1" ]] && echo -e "${NC}[INFO]:    $*" && echo -ne "${NC}"
 }
 
-function askInstallMasternode(){
+function askInstallDefiMasternode(){
 
 	if [ -d "$MASTERNODE_TARGET_FOLDER" ]; 
 		then #Masternode-folder already exists?
 		echo ''
 			while true; do
-
-			case $(promptValue "Masternode-folder already exists, continue and overwrite?(y/n)" "y") in 
+			case $(promptValue "Masternode-folder already exists, continue installation and overwrite?(y/n)" "y") in 
 			  y|Y )
-			   
-                installMasternode
+			          rm -r "${MASTERNODE_TARGET_FOLDER:?}"
+                installDefiMasternode
 				
 			    break ;;
 			  n|N )
-					__msg_warning "Installation skipped by user"
+					__msg_warning "Masternode installation skipped by user"
 
-                    exit ${INSTALLATION_ABORTED}
+                    exit ${EXIT_CODE_INSTALLATION_ABORTED}
 					break ;;
 			  * ) 	__msg_warning "Invalid input"
 					echo -e "${NC}";;
 			esac
 	    done
 	else
-		installMasternode
+		installDefiMasternode
 	fi
 }
 
-function installMasternode(){
-    wget -P /tmp $MASTERNODE_DOWNLOAD_URL
-    tar -xvzf /tmp/$MASTERNODE_ARCHIVE -C /tmp/
+function installDefiMasternode(){
+
+    __msg_info "Install Masternode"
+
+    wget -P /tmp -q --show-progress $MASTERNODE_DOWNLOAD_URL
+    tar -xzf /tmp/$MASTERNODE_ARCHIVE -C /tmp/
     mkdir -p "$MASTERNODE_TARGET_FOLDER"
-    rm -r "${MASTERNODE_TARGET_FOLDER:?}/"*
     cp -R /tmp/$MASTERNODE_RELEASE_FOLDER/* "$MASTERNODE_TARGET_FOLDER"
+
+    chown -R $USERNAME: "$MASTERNODE_TARGET_FOLDER"
+
+    __msg_success "Masternode installed"
+}
+
+function downloadDefiSnapshot(){
+
+    __msg_info "Install snapshot"
+
+    mkdir -p /tmp/snapshot
+    wget -P /tmp/snapshot -q --show-progress $SNAPSHOT_DOWNLOAD_URL
+    unzip -q /tmp/snapshot/*.zip -d /tmp/snapshot/
+    rm -Rf "${MASTERNODE_TARGET_FOLDER:?}/"chainstate 
+    rm -Rf "${MASTERNODE_TARGET_FOLDER:?}/"enhancedcs
+    rm -Rf "${MASTERNODE_TARGET_FOLDER:?}/"blocks
+
+    mv /tmp/snapshot/* "$MASTERNODE_TARGET_FOLDER"
+
+    chown -R $USERNAME: "$MASTERNODE_TARGET_FOLDER"
+
+    __msg_success "Snapshot installed"
+}
+
+function changesshPort(){
+
+    # We dice ourselves a new SSH port and change the config
+    SSH_PORT=$(( ((RANDOM<<15)|RANDOM) % 63001 + 2000 ))
+    sed -i '/^#Port/s/#Port/Port/' /etc/ssh/sshd_config
+    sed -i "/^Port/s/22/${SSH_PORT}/g" /etc/ssh/sshd_config
+    sed -i '/^PermitRootLogin/s/yes/no/' /etc/ssh/sshd_config
+    echo "ATTENTION: use the port $SSH_PORT at the next LOGIN via SSH. Ex: ssh $USERNAME@yourIP -p $SSH_PORT."
 }
 
 #_start________________________________________________________________________
 
 if [ "$EUID" -ne 0 ]
-  then echo "Please execute as root"
-  exit ${NO_ROOT}
+  then __msg_error "Please execute as root"
+  exit ${EXIT_CODE_NO_ROOT}
 fi
 
 clear
 
-askInstallMasternode
+USERNAME=$(promptValue "Please select the username that should run the masternode process. Enter for default name" $USERNAME)
+MASTERNODE_TARGET_FOLDER="/home/$USERNAME/.defi"
+
+# create user to run the masternode
+id -u "$USERNAME" &>/dev/null || adduser --gecos "" "$USERNAME"
+
+__msg_info "System is being updated"
+DEBIAN_FRONTEND=noninteractive apt-get update -qq < /dev/null > /dev/null
+DEBIAN_FRONTEND=noninteractive apt-get upgrade -qq < /dev/null > /dev/null
+DEBIAN_FRONTEND=noninteractive apt-get install -qq ufw nano htop fail2ban psmisc unzip wget < /dev/null > /dev/null
+
+askInstallDefiMasternode
+#downloadDefiSnapshot
+
+changesshPort
+
+# enter passwords and data for user
+
+# Allow required ports in the firewall
+ufw allow $SSH_PORT/tcp
+ufw allow 8555/tcp
+
+# Restart SSH
+systemctl restart ssh
+
+# Enable firewall
+ufw enable
+
+# Start defi daemon
+"${MASTERNODE_TARGET_FOLDER:?}/"bin/defid -daemon
+
+
+
+
